Set the default value for (db-inquirer)'s `default' keyword param as '()
To be consistent with other db-support eggs for awful.

WARNING: this change may break your code!  In case of an empty result
set from a query, (db-inquirer) will now return '() instead of #f.

This change can be specially harmful in case you have something like

    (or ($db "some query") "something")

In this example, if the query returns an empty result set, that
expression would produce:

* in older awful-postgresql versions: "something"
* in awful-postgresql versions with this change: '()